Restructure routing tests
