+++ tests/lib/Integration/Routing/CustomTest.php
@@ -1,14 +1,30 @@
 <?php
+/**
+ * Copyright 2019 Cloud Creativity Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-namespace CloudCreativity\LaravelJsonApi\Tests\Integration;
+namespace CloudCreativity\LaravelJsonApi\Tests\Integration\Routing;
 
 use CloudCreativity\LaravelJsonApi\Routing\RouteRegistrar;
 use DummyApp\Comment;
 use DummyApp\Jobs\SharePost;
 use DummyApp\Post;
 use Illuminate\Support\Facades\Queue;
+use CloudCreativity\LaravelJsonApi\Tests\Integration\TestCase;
 
-class CustomRoutesTest extends TestCase
+class CustomTest extends TestCase
 {
 
     /**
